+++ src/resources/offer.rs
@@ -1,13 +1,15 @@
 use crate::link::Link;
-use crate::resources::{Asset, Price};
+use crate::resources::{Asset, LedgerId, OfferId, Price};
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
+use serde_with::rust::display_fromstr;
 
 #[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
 pub struct Offer {
     #[serde(rename = "_links")]
     pub links: OfferLinks,
-    pub id: String,
+    #[serde(with = "display_fromstr")]
+    pub id: OfferId,
     pub paging_token: String,
     pub seller: String,
     pub selling: Asset,
@@ -16,7 +18,7 @@ pub struct Offer {
     #[serde(rename = "price_r")]
     pub price_ratio: Price,
     pub price: String,
-    pub last_modified_ledger: i32,
+    pub last_modified_ledger: LedgerId,
     pub last_modified_time: Option<DateTime<Utc>>,
 }